development-erp/create-erp-development-db-tables-app/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.37
PyMySQL==1.1.1
pandas==2.2.3
numpy==2.2.1
pydantic-settings==2.7.1
cryptography==44.0.0
